Don't cast away provenance in ptr2int casts
When casting from a pointer to an integer that can hold a pointer (e.g.,
uint64_t in the current code base), cast to uintptr_t first and allow an
implicit cast to make the final cast rather than doing:

 a_uint64 = (uint64_t)(uintptr_t)a_pointer;

This will avoid loss of provenance on CHERI when the destination is
updated to be a type capable of holding a capability.

+++ module/nvpair/nvpair.c
@@ -553,7 +553,7 @@ nvlist_init(nvlist_t *nvl, uint32_t nvflag, nvpriv_t *priv)
 {
 	nvl->nvl_version = NV_VERSION;
 	nvl->nvl_nvflag = nvflag & (NV_UNIQUE_NAME|NV_UNIQUE_NAME_TYPE);
-	nvl->nvl_priv = (uint64_t)(uintptr_t)priv;
+	nvl->nvl_priv = (uintptr_t)priv;
 	nvl->nvl_flag = 0;
 	nvl->nvl_pad = 0;
 }
